updated helper function directories
