+++ front-end/src/layouts/AdminDashboard/AdminDashboard.js
@@ -1,5 +1,12 @@
 import { useState, useEffect, useRef } from 'react';
 import './AdminDashboard.css';
+import SearchBarAdmin from '../../components/admin/SearchBarAdmin/SearchBarAdmin';
+import {
+    sortByPriorityAndDate,
+    sortByPriorityReverse,
+    sortByDateAscending,
+    sortByDate
+} from '../../components/admin/helper/sorting/SortingFunctions';
 
 function AdminDashboard() {
     const [searchText, setSearchText] = useState('');
@@ -7,6 +14,7 @@ function AdminDashboard() {
     const [activeOptionsOverlay, setActiveOptionsOverlay] = useState(null);
     const [isOverlayOptionsOpen, setIsOverlayOptionsOpen] = useState(false);
     const [columnSortOptions, setColumnSortOptions] = useState({});
+    const overlayRef = useRef(null);
     const currentDepartment = "IT"; // will change this later sprint
 
     useEffect(() => {
@@ -16,6 +24,31 @@ function AdminDashboard() {
             .then(data => setIssues(data))
             .catch(error => console.error('Error fetching data:', error));
     }, []);
+
+    // Add event listener to handle clicks outside the overlay
+    useEffect(() => {
+        const handleOutsideClick = (e) => {
+            if (overlayRef.current && !overlayRef.current.contains(e.target)) {
+                closeOverlayOptions();
+            }
+        };
+
+        if (isOverlayOptionsOpen) {
+            document.addEventListener('mousedown', handleOutsideClick);
+        } else {
+            document.removeEventListener('mousedown', handleOutsideClick);
+        }
+
+        return () => {
+            document.removeEventListener('mousedown', handleOutsideClick);
+        };
+    }, [isOverlayOptionsOpen]);
+
+    // Function to close the overlay
+    const closeOverlayOptions = () => {
+        setActiveOptionsOverlay(null);
+        setIsOverlayOptionsOpen(false);
+    };
 
     // Filter issues based on the search text
     const filteredIssues = issues.filter(issue => {
@@ -47,41 +80,7 @@ function AdminDashboard() {
         'Action Required': 'Awaiting Response',
         'Resolved': 'Resolved'
     };
-
-    const priorityOrder = {
-        'Reopened': 1,
-        'High Priority': 2,
-        'New': 3
-    };
     /* eslint-enable quote-props */
-
-    // Custom date parsing function for "dd/mm/yyyy" format
-    const parseDate = (dateString) => {
-        const [day, month, year] = dateString.split('/').map(Number);
-        return new Date(year, month - 1, day); // Month is 0-based in JavaScript
-    };
-
-    const sortByPriorityAndDate = (a, b) => {
-        if (a.currentPriority === b.currentPriority) {
-            return parseDate(b.dateCreated) - parseDate(a.dateCreated);
-        }
-        return priorityOrder[a.currentPriority] - priorityOrder[b.currentPriority];
-    };
-
-    const sortByPriorityReverse = (a, b) => {
-        if (a.currentPriority === b.currentPriority) {
-            return parseDate(b.dateCreated) - parseDate(a.dateCreated);
-        }
-        return priorityOrder[b.currentPriority] - priorityOrder[a.currentPriority];
-    };
-
-    const sortByDateAscending = (a, b) => {
-        return parseDate(a.dateCreated) - parseDate(b.dateCreated);
-    };
-
-    const sortByDate = (a, b) => {
-        return parseDate(b.dateCreated) - parseDate(a.dateCreated);
-    };
 
     const getSortingFunctionForColumn = (status) => {
         switch (columnSortOptions[status]) {
@@ -103,47 +102,10 @@ function AdminDashboard() {
             .sort(getSortingFunctionForColumn(newStatus))
     }));
 
-    // Ref for the overlay div
-    const overlayRef = useRef(null);
-
-    // Function to close the overlay
-    const closeOverlayOptions = () => {
-        setActiveOptionsOverlay(null);
-        setIsOverlayOptionsOpen(false);
-    };
-
-    // Add event listener to handle clicks outside the overlay
-    useEffect(() => {
-        const handleOutsideClick = (e) => {
-            if (overlayRef.current && !overlayRef.current.contains(e.target)) {
-                closeOverlayOptions();
-            }
-        };
-
-        if (isOverlayOptionsOpen) {
-            document.addEventListener('mousedown', handleOutsideClick);
-        } else {
-            document.removeEventListener('mousedown', handleOutsideClick);
-        }
-
-        return () => {
-            document.removeEventListener('mousedown', handleOutsideClick);
-        };
-    }, [isOverlayOptionsOpen]);
-
     return (
         <div className="admin-dashboard">
             <h1>Issue Board</h1>
-            <div className="search-bar-admin">
-                <form onSubmit={(e) => e.preventDefault()}>
-                    <input
-                        type="text"
-                        placeholder="Search by keyword, netid:, name:, date:, priority:"
-                        value={searchText}
-                        onChange={(e) => setSearchText(e.target.value)}
-                    />
-                </form>
-            </div>
+            <SearchBarAdmin searchText={searchText} onSearchTextChange={setSearchText} />
             <div className="issue-columns">
                 {groupedAndOrderedIssues.map(({ status, issues }) => (
                     <div key={status} className="issue-column">
